Extract the Codex glue to lib/runtime/codex-glue.js
Verbatim move of the Codex runtime span: the shared app-server lifecycle
(lazy singleton with module-owned state), direct-chat turns, delegate turn
wiring, approvals routing, the keepalive heartbeat, and every failure
surface. Mechanical conversions only:

- Workspace root read at USE time via lib/config.js (6 sites): a workspace
  switch immediately redirects thread cwds and instruction reads.
- Former root deps that are now lib modules are required directly:
  recordEvent (signals), resolveMarkers (delegation/markers),
  buildSystemPrompt (agents/prompt), readNormalisedFile (agents/discovery).
- Root-owned capabilities arrive through wireCodexGlueDeps with throwing
  unwired defaults, wired at boot: chatProcesses and recentSpawnErrors as
  identity accessors (the glue mutates the root's own maps), safeSend,
  appendTranscript, endConvoTransition, child-pid registry,
  killProcessTree, requestServerPermission, and getters for the listen
  port and permission timeout.

What deliberately stayed in the root: getRuntimeStatus plus the Claude
probe caches (settings machinery that sat mid-span), requestServerPermission
(the generic server-originated permission bridge, used beyond Codex;
injected), and the kill-window machine. The root's synchronous exit
handler reads the app-server pid through a glue accessor (a primitive
cannot be re-exported by identity).

runtime-adapter.test.js pins retargeted in the same commit so every commit
stays green: the shared-server and turn seams (and the approval-bridge use)
now pin the glue source; spawn/routing/bridge-definition pins stay on
server.js. The unused codex-appserver require leaves the root.

server.js: 6,450 -> 5,573 lines.
