+++ test/unit/runtime-adapter.test.js
@@ -13,6 +13,7 @@ const path = require('node:path');
 
 const ROOT = path.join(__dirname, '..', '..');
 const serverSrc = fs.readFileSync(path.join(ROOT, 'server.js'), 'utf-8');
+const glueSrc = fs.readFileSync(path.join(ROOT, 'lib', 'runtime', 'codex-glue.js'), 'utf-8');
 const docSrc = fs.readFileSync(path.join(ROOT, 'docs', 'RUNTIME-ADAPTER.md'), 'utf-8');
 
 describe('runtime adapter contract: the documented seams exist', () => {
@@ -36,14 +37,14 @@ describe('runtime adapter contract: the documented seams exist', () => {
 
   test('the server routes both runtimes through the documented spawn seams', () => {
     assert.match(serverSrc, /function spawnClaude/, 'Claude spawn seam');
-    assert.match(serverSrc, /function getCodexAppServer/, 'Codex shared-server seam');
-    assert.match(serverSrc, /function startCodexTurn/, 'Codex turn seam');
+    assert.match(glueSrc, /function getCodexAppServer/, 'Codex shared-server seam (lib/runtime/codex-glue.js)');
+    assert.match(glueSrc, /function startCodexTurn/, 'Codex turn seam (lib/runtime/codex-glue.js)');
     assert.match(serverSrc, /runtime\s*===\s*'codex'|runtime:\s*'codex'/, 'runtime routing by frontmatter field');
   });
 
   test('approvals route through the shared permission bridge', () => {
-    assert.match(serverSrc, /function requestServerPermission/, 'server-originated permission bridge');
-    assert.match(serverSrc, /handleCodexApproval|requestServerPermission\(/, 'Codex approvals use the bridge');
+    assert.match(serverSrc, /function requestServerPermission/, 'server-originated permission bridge (stays in the root: used beyond Codex)');
+    assert.match(glueSrc, /handleCodexApproval|requestServerPermission\(/, 'Codex approvals use the injected bridge');
   });
 
   test('the contract document names every seam it pins', () => {
